Cleaned up index.js server file
